TNT in claimed areas
There are still some loopholes (just because the code doesn't stop it doesn't mean it's allowed on the server) but it generally works.
- Players cannot ignite TNT in others' claims, including both direct interaction and flame bows.
- TNT will not destroy blocks in spawn or in claims outside of their origin claim (either where the TNT block was, or the dispenser block that placed it)
Basically, the owner is set from when the TNT is ignited, and it can only modify what that player can modify. Players should not be able to trick the plugin into setting a different player as the owner of the TNT.

- Switched dependency to Yatopia for access to Paper API